requirements.txt
aiogram==3.21.0
aiosqlite
python-dotenv
